Creation of a new diodeSimulator class in utilities to lighten up and rationalise a bit the ARA trigger code. The functions in the triggerSimulator have already changed to try to minimise user-induced mistakes.
