+++ NuRadioReco/modules/ARA/triggerSimulator.py
@@ -6,6 +6,7 @@ import logging
 import scipy.signal
 from NuRadioReco.modules.channelGenericNoiseAdder import channelGenericNoiseAdder
 from NuRadioReco.framework.trigger import IntegratedPowerTrigger
+from NuRadioReco.utilities.diodeSimulator import diodeSimulator
 import NuRadioReco.framework.channel
 
 logger = logging.getLogger('ARAtriggerSimulator')
@@ -15,149 +16,15 @@ class triggerSimulator:
     """
     Calculates the trigger of an event.
     Uses the ARA trigger logic of a tunnel diode.
-    Implementation as in PyRex by Ben Hokanson-Fasig/
+    Implementation similar to PyRex by Ben Hokanson-Fasig/
     """
 
     def __init__(self):
         self.__t = 0
-        self.begin()
+        self._power_mean = None
+        self._power_std = None
+        self._diode = diodeSimulator()
         logger.warning("This module does not contain cutting the trace to ARA specific parameters.")
-
-    def begin(self, antenna_resistance=8.5 * units.ohm,
-              power_mean=None,
-              power_std=None):
-        """
-        Calculate a signal as processed by the tunnel diode.
-        The given signal is convolved with the tunnel diodde response as in
-        AraSim.
-
-        Parameters
-        ----------
-        antenna_resistance : float
-            Value of the resistance of the ARA antennas
-        """
-
-        self.antenna_resistance = antenna_resistance
-        self._power_mean = power_mean
-        self._power_std = power_std
-
-    # Tunnel diode response functions pulled from arasim
-    # RL (Robert Lahmann) Sept 3, 2018: this is not documented in the arasim code, but it seems most
-    # logical to assume that the units of the two middle parameters are seconds and that the
-    # other parameters are unitless
-    _td_args = {
-        'down1': (-0.8, 15e-9 * units.s, 2.3e-9 * units.s, 0),
-        'down2': (-0.2, 15e-9 * units.s, 4e-9 * units.s, 0),
-        'up': (1, 18e-9 * units.s, 7e-9 * units.s, 1e9)
-    }
-    # Set td_args['up'][0] based on the other args, like in arasim
-    _td_args['up'] = (-np.sqrt(2 * np.pi) *
-                      (_td_args['down1'][0] * _td_args['down1'][2] +
-                       _td_args['down2'][0] * _td_args['down2'][2]) /
-                      (2e18 * _td_args['up'][2] ** 3),) + _td_args['up'][1:]
-
-    # Set "down" and "up" functions as in arasim
-    @classmethod
-    def _td_fdown1(cls, x):
-        return (cls._td_args['down1'][3] + cls._td_args['down1'][0] *
-                np.exp(-(x - cls._td_args['down1'][1]) ** 2 /
-                       (2 * cls._td_args['down1'][2] ** 2)))
-
-    @classmethod
-    def _td_fdown2(cls, x):
-        return (cls._td_args['down2'][3] + cls._td_args['down2'][0] *
-                np.exp(-(x - cls._td_args['down2'][1]) ** 2 /
-                       (2 * cls._td_args['down2'][2] ** 2)))
-
-    @classmethod
-    def _td_fup(cls, x):
-        return (cls._td_args['up'][0] *
-                (cls._td_args['up'][3] * (x - cls._td_args['up'][1])) ** 2 *
-                np.exp(-(x - cls._td_args['up'][1]) / cls._td_args['up'][2]))
-
-    def tunnel_diode(self, channel):
-        """
-        Calculate a signal as processed by the tunnel diode.
-        The given signal is convolved with the tunnel diodde response as in
-        AraSim.
-        Parameters
-        ----------
-        channel : Channel
-            Signal to be processed by the tunnel diode.
-        power_mean : float
-            Parameter extracted in ARA from noise.
-            If not given, it is calculated from generic noise
-        power_std : float
-            Parameter extracted in ARA from noise.
-            If not given, it is calculated from generic noise
-
-        Returns
-        -------
-        trace_after_tunnel_diode: array
-            Signal output of the tunnel diode for the input `channel`.
-
-        """
-        t_max = 1e-7 * units.s
-        n_pts = int(t_max * channel.get_sampling_rate())
-        times = np.linspace(0, t_max, n_pts + 1)
-        diode_resp = self._td_fdown1(times) + self._td_fdown2(times)
-        t_slice = times > self._td_args['up'][1]
-        diode_resp[t_slice] += self._td_fup(times[t_slice])
-        conv = scipy.signal.convolve(channel.get_trace() ** 2 / self.antenna_resistance,
-                                     diode_resp, mode='full')
-        # conv multiplied by dt so that the amplitude stays constant for
-        # varying dts (determined emperically, see ARVZAskaryanSignal comments)
-        # Setting output
-        trace_after_tunnel_diode = conv / channel.get_sampling_rate()
-        trace_after_tunnel_diode = trace_after_tunnel_diode[:channel.get_trace().shape[0]]
-
-        return trace_after_tunnel_diode
-
-    def calculate_noise_parameters(self,
-                                   sampling_rate = 1 * units.GHz,
-                                   min_freq = 50 * units.MHz,
-                                   max_freq = 1 * units.GHz,
-                                   amplitude = 10 * units.microvolt,
-                                   type='rayleigh'):
-        """
-        Calculates the mean and the standard deviation for the diode-filtered noise.
-
-        Parameters
-        ----------
-        sampling_rate: float
-            Sampling rate
-        min_freq: float
-            Minimum frequency of the bandwidth
-        max_freq: float
-            Maximum frequency of the bandwidth
-        amplitude: float
-            Voltage amplitude (RMS) for the noise
-        type: string
-            Noise type
-
-        Returns
-        -------
-        power_mean: float
-            Mean of the diode-filtered noise
-        power_std: float
-            Standard deviation of the diode-filtered noise
-        """
-        noise = NuRadioReco.framework.channel.Channel(0)
-
-        long_noise = channelGenericNoiseAdder().bandlimited_noise(min_freq=min_freq,
-                                        max_freq=max_freq,
-                                        n_samples=10000,
-                                        sampling_rate=sampling_rate,
-                                        amplitude=amplitude,
-                                        type=type)
-
-        noise.set_trace(long_noise, sampling_rate)
-        power_noise = self.tunnel_diode(noise)
-
-        power_mean = np.mean(power_noise)
-        power_std = np.std(power_noise)
-
-        return power_mean, power_std
 
     def has_triggered(self, channel):
         """
@@ -175,31 +42,9 @@ class triggerSimulator:
         boolean
             Whether or not the antenna triggers on `channel`.
         """
-        if self._power_mean is None or self._power_std is None:
-            # Prepare for antenna trigger by finding rms of noise waveform
-            # (1 microsecond) convolved with tunnel diode response
-
-            # This is not fully true yet, since we don't have ARA frontend implemeted
-            # long_noise is therefore just set to a certain value rather
-            # than taken the full ARA signal chain
-            noise = NuRadioReco.framework.channel.Channel(0)
-
-            long_noise = channelGenericNoiseAdder().bandlimited_noise(min_freq=50 * units.MHz,
-                                            max_freq=1000 * units.MHz,
-                                            n_samples=10000,
-                                            sampling_rate=channel.get_sampling_rate(),
-                                            amplitude=20 * units.mV,
-                                            type='perfect_white')
-
-            noise.set_trace(long_noise, channel.get_sampling_rate())
-
-            self.__power_noise = self.tunnel_diode(noise)
-
-            self._power_mean = np.mean(self.__power_noise)
-            self._power_std = np.std(self.__power_noise)
 
         # Send signal through tunnel_diode
-        after_tunnel_diode = self.tunnel_diode(channel)
+        after_tunnel_diode = self._diode.tunnel_diode(channel)
         low_trigger = (self._power_mean -
                        self._power_std * np.abs(self.power_threshold))
 
@@ -242,6 +87,11 @@ class triggerSimulator:
         if(power_mean is not None and power_std is not None):
             self._power_mean = power_mean
             self._power_std = power_std
+        else:
+            error_msg  = 'The power_mean and power_std parameters are not defined. '
+            error_msg += 'Please define them. You can use the calculate_noise_parameters '
+            error_msg += 'function in utilities.diodeSimulator to do so.'
+            raise ValueError(error_msg)
 
         self.power_threshold = power_threshold
 
@@ -252,7 +102,7 @@ class triggerSimulator:
         times_max = []
         sampling_rates = []
         number_triggered_channels = 0
-        
+
         for channel in station.iter_channels():
             channel_id = channel.get_id()
             if channel_id not in triggered_channels:
@@ -261,7 +111,7 @@ class triggerSimulator:
             if trigger[channel_id]:
                 number_triggered_channels += 1
                 times = channel.get_times()
-                trace_after_diode = self.tunnel_diode(channel)
+                trace_after_diode = self._diode.tunnel_diode(channel)
                 arg_trigger = np.argmin(trace_after_diode)
                 trigger_times.append(times[arg_trigger])
                 times_min.append(np.min(times))
